// libs/report-states/ui/src/lib/adapters/reject-report-states-vm-props.interface.ts
import { TableRowBase } from '@cmz/shared-ui';

export interface RejectReportStatesVmProps extends TableRowBase {
    uniqId: string;
    reportTypeLabel: string;
    operatorsLabel: string;
    sourceLabel: string;
    initiatorPhoneNumber: string;
    reportedAt: string;
    actionsRef: string;
}